Added Exception and Validation

+++ src/main/java/com/arsinha/springlearning/controller/StudentController.java
@@ -1,9 +1,13 @@
 package com.arsinha.springlearning.controller;
 
 import java.util.ArrayList;
-import java.util.List;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,6 +16,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.arsinha.springlearning.exceptions.StudentNotFoundException;
 import com.arsinha.springlearning.model.Student;
 import com.arsinha.springlearning.service.StudentService;
 
@@ -30,21 +35,49 @@ public class StudentController {
 	}
 	
 	@PostMapping("/student")
-	private void CreateStudent(@RequestBody Student student) {
+	private ResponseEntity<?> createStudent(@Valid @RequestBody Student student) {
+		
 		studentService.saveOrUpdate(student);
+		return new ResponseEntity<>(HttpStatus.CREATED);
 	}
 	@PutMapping("/student")
-	private void UpdateStudent(@RequestBody Student student) {
+	private ResponseEntity<?> updateStudent(@Valid @RequestBody Student student) {
 		studentService.saveOrUpdate(student);
+		return new ResponseEntity<>(student, HttpStatus.OK);
 	}
 	@GetMapping("/student/{studentId}")
-	private Student getStudentById(@PathVariable("studentId") int studentId) {
+	private Student getStudentById(@PathVariable("studentId") int studentId) throws Exception {
+		try {
 		return studentService.getStudentById(studentId);
+		}
+		catch(InvalidDataAccessApiUsageException e) {
+			throw new StudentNotFoundException("studentId: "+ studentId);
+
+		}
+		catch(Exception ex) {
+			throw new Exception("studentId: "+ studentId);
+		}
 	}
 
 	@DeleteMapping("/student/{studentId}")
-	private void deleteStudentById(@PathVariable("studentId") int studentId) {
-		 studentService.deleteStudentById(studentId);
-	}
+	private void deleteStudentById(@PathVariable("studentId") int studentId) throws Exception {
+		try{Student student=studentService.getStudentById(studentId);
+		 if(student==null) { 
+			//runtime exception  
+			throw new StudentNotFoundException("studentId: "+ studentId);
+		 }
+		 else {
+			 studentService.deleteStudentById(studentId);
+		 }
+		}catch(InvalidDataAccessApiUsageException e) {
+			throw new StudentNotFoundException("studentId: "+ studentId);
+
+		}
+		catch(Exception ex) {
+			ex.printStackTrace();
+			throw new Exception("studentId: "+ studentId);
+		}
+			}  
+	
 
 }
